+++ src/types/index.ts
@@ -60,3 +60,4 @@ export * from './databaseTypes';
 export * from './userTypes';
 export * from './authTypes';
 export * from './registrationTypes';
+export * from './validationTypes';
