+++ src/types/registration.ts
@@ -1,3 +1,5 @@
+import { ValidationErrors } from './validationTypes';
+
 /**
  * Registration form data structure for user sign-up
  */
@@ -14,52 +16,14 @@ export interface RegistrationFormData {
 }
 
 /**
- * Validation errors for registration form
- */
-export interface RegistrationErrors {
-  fullName?: string;
-  email?: string;
-  confirmEmail?: string;
-  password?: string;
-  confirmPassword?: string;
-  displayName?: string;
-  sportsInterested?: string;
-  skillLevel?: string;
-  zipCode?: string;
-  general?: string;
-}
-
-/**
- * ZIP code lookup response
- */
-export interface ZipCodeResponse {
-  city: string;
-  state: string;
-  fullLocation: string;
-}
-
-/**
- * Registration validation result
- */
-export interface RegistrationValidationResult {
-  isValid: boolean;
-  errors: RegistrationErrors;
-  cityInfo?: ZipCodeResponse;
-}
-
-/**
  * Registration API response
  */
 export interface RegistrationResponse {
   success: boolean;
   message?: string;
   error?: string;
-  data?: {
-    user: {
-      uid: string;
-      email: string;
-      displayName: string;
-      cityInfo?: ZipCodeResponse;
-    };
-  };
-} 
+  validationErrors?: ValidationErrors;
+}
+
+// Re-export for backward compatibility
+export type RegistrationErrors = ValidationErrors;
