+++ src/types/registrationTypes.ts
@@ -1,3 +1,5 @@
+import { ValidationErrors } from './validationTypes';
+
 /**
  * Registration form data structure for user sign-up
  */
@@ -14,38 +16,14 @@ export interface RegistrationFormData {
 }
 
 /**
- * Validation errors for registration form
- */
-export interface RegistrationErrors {
-  fullName?: string;
-  email?: string;
-  confirmEmail?: string;
-  password?: string;
-  confirmPassword?: string;
-  displayName?: string;
-  sportsInterested?: string;
-  skillLevel?: string;
-  zipCode?: string;
-  general?: string;
-}
-
-/**
  * Registration API response
  */
 export interface RegistrationResponse {
   success: boolean;
   message?: string;
   error?: string;
-  validationErrors?: {
-    email?: string;
-    password?: string;
-    fullName?: string;
-    displayName?: string;
-    sportsInterested?: string;
-    skillLevel?: string;
-    zipCode?: string;
-    confirmEmail?: string;
-    confirmPassword?: string;
-    general?: string;
-  };
-} 
+  validationErrors?: ValidationErrors;
+}
+
+// Re-export for backward compatibility
+export type RegistrationErrors = ValidationErrors;
